requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.5.2
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
progress==1.6
propcache==0.3.1
python-dotenv==1.1.0
requests==2.32.3
sans==1.2.6
sniffio==1.3.1
tzlocal==5.3.1
urllib3==2.4.0
yarl==1.20.0
